chore: fix linter issues & clidoc CI
