+++ internal/relationtuple/definitions.go
@@ -65,33 +65,6 @@ type RelationQuery struct {
 	SubjectSet *SubjectSet `json:"subject_set,omitempty"`
 }
 
-// swagger:model InternalRelationTuple
-type relationTupleWithRequired struct {
-	// Namespace of the Relation Tuple
-	//
-	// required: true
-	Namespace string `json:"namespace"`
-
-	// Object of the Relation Tuple
-	//
-	// required: true
-	Object string `json:"object"`
-
-	// Relation of the Relation Tuple
-	//
-	// required: true
-	Relation string `json:"relation"`
-
-	// SubjectID of the Relation Tuple
-	//
-	// Either SubjectSet or SubjectID are required.
-	SubjectID *string `json:"subject_id,omitempty"`
-	// SubjectSet of the Relation Tuple
-	//
-	// Either SubjectSet or SubjectID are required.
-	SubjectSet *SubjectSet `json:"subject_set,omitempty"`
-}
-
 // swagger:ignore
 type TupleData interface {
 	// swagger:ignore
@@ -161,20 +134,6 @@ const (
 	ActionInsert patchAction = "insert"
 	ActionDelete patchAction = "delete"
 )
-
-// The patch request payload
-//
-// swagger:parameters patchRelationTuples
-// nolint:deadcode,unused
-type patchPayload struct {
-	// in:body
-	Payload []*PatchDelta
-}
-
-type PatchDelta struct {
-	Action        patchAction            `json:"action"`
-	RelationTuple *InternalRelationTuple `json:"relation_tuple"`
-}
 
 func SubjectFromString(s string) (Subject, error) {
 	if strings.Contains(s, "#") {
